requirements.txt
asgiref==3.4.1
boto3==1.5.36
botocore==1.8.50
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.1.7
django-bootstrap-form==3.3
django-debug-toolbar==3.2.1
django-storages==1.6.5
docutils==0.14
gunicorn==20.1.0
jmespath==0.9.3
Pillow==8.3.0
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.3
s3transfer==0.1.13
six==1.11.0
sqlparse==0.4.1
static3==0.7.0
